requirements.txt
Flask == 1.1.2
PyJWT == 1.7.1
SQLAlchemy == 1.3.19